# api/openapi.yml
components:
  responses:
    HTTPError:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
      description: Error response
  schemas:
    HTTPError:
      additionalProperties: true
      properties:
        code:
          type: string
        error:
          type: string
      required:
      - code
      - error
      type: object
    handler_SandboxDetailResponse:
      properties:
        api_key_id:
          description: APIKeyID is the ID of the API key that created the sandbox.
          type: string
        client_id:
          description: ClientID identifies this platform to SDK clients.
          type: string
        end_at:
          description: EndAt is the scheduled or actual sandbox end time.
          format: date-time
          nullable: true
          type: string
        envd_version:
          description: EnvdVersion is the envd component version inside the sandbox.
          type: string
        sandbox_id:
          description: SandboxID is the sandbox instance ID.
          type: string
        started_at:
          description: StartedAt is the sandbox start time.
          format: date-time
          nullable: true
          type: string
        state:
          description: State is the current sandbox state.
          type: string
        template_id:
          description: TemplateID is the template used by the sandbox.
          type: string
      type: object
    handler_SandboxLogsResponse:
      properties:
        log_entries:
          description: LogEntries is the structured log entry list field.
          items:
            $ref: "#/components/schemas/service_SandboxLogEntry"
          type: array
        logs:
          description: Logs is the legacy SDK-compatible log list field.
          items:
            $ref: "#/components/schemas/service_SandboxLogEntry"
          type: array
      type: object
...
    handler_SandboxNetworkRouteConfig:
      properties:
        allow_out:
          description: AllowOut lists outbound targets that are explicitly allowed.
          items:
            type: string
          type: array
        allow_public_traffic:
          description: AllowPublicTraffic controls whether public traffic may reach the sandbox.
          type: boolean
        deny_out:
          description: DenyOut lists outbound targets that are explicitly denied.
          items:
            type: string
          type: array
        mask_request_host:
          description: MaskRequestHost specifies the request host to mask or rewrite.
          type: string
      type: object
...
    handler_UpdateTemplateResponse:
      properties:
        names:
          description: Names returns the updated template name list.
          items:
            type: string
          type: array
      type: object
...
info:
  title: AoneSuite Infra API
  version: 0.0.0
openapi: 3.0.3
paths:
...
  /api/v1/sbx/sandboxes:
    get:
      description: ListSandboxes handles `GET /api/v1/sbx/sandboxes`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_ListSandboxes
      parameters:
      - in: query
        name: api_key_id
        schema:
          description: APIKeyID filters by the API key that created the resource; empty means no API key filter.
          type: string
      - in: query
        name: template_id
        schema:
          description: TemplateID filters sandbox instances by template ID.
          type: string
      - in: query
        name: state
        schema:
          description: State filters sandbox instances by a comma-separated list of states.
          type: string
      - in: query
        name: cursor
        schema:
          description: Cursor is the pagination cursor.
          type: string
      - in: query
        name: limit
        schema:
          description: Limit is the maximum number of items to return.
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/handler_SandboxDetailResponse"
                type: array
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: ListSandboxes handles `GET /api/v1/sbx/sandboxes`.
    post:
      description: CreateSandbox handles `POST /api/v1/sbx/sandboxes`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_CreateSandbox
      requestBody:
        content:
          application/json:
            schema:
              properties:
                allow_internet_access:
                  description: AllowInternetAccess controls public internet access; nil uses the provider default.
                  type: boolean
                env_vars:
                  additionalProperties:
                    type: string
                  description: EnvVars stores environment variables injected when creating the sandbox.
                  type: object
                metadata:
                  additionalProperties:
                    type: string
                  description: Metadata stores caller-provided sandbox metadata.
                  type: object
                network:
                  $ref: "#/components/schemas/handler_SandboxNetworkRouteConfig"
                secure:
                  description: Secure controls whether provider-specific secure isolation is enabled.
                  type: boolean
                template_id:
                  description: TemplateID selects the template used to create the sandbox.
                  type: string
                timeout:
                  description: Timeout sets the sandbox lifetime in seconds; 0 uses the provider default.
                  format: int32
                  type: integer
              type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_SandboxResponse"
          description: Created
        default:
          $ref: "#/components/responses/HTTPError"
      summary: CreateSandbox handles `POST /api/v1/sbx/sandboxes`.
  /api/v1/sbx/sandboxes/{sandbox_id}:
    delete:
      description: DeleteSandbox handles `DELETE /api/v1/sbx/sandboxes/{sandbox_id}`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_DeleteSandbox
      parameters:
      - in: path
        name: sandbox_id
        required: true
        schema:
          description: SandboxID is bound from the `:sandbox_id` route parameter.
          type: string
      responses:
        "200":
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: DeleteSandbox handles `DELETE /api/v1/sbx/sandboxes/{sandbox_id}`.
    get:
      description: GetSandbox handles `GET /api/v1/sbx/sandboxes/{sandbox_id}`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_GetSandbox
      parameters:
      - in: path
        name: sandbox_id
        required: true
        schema:
          description: SandboxID is bound from the `:sandbox_id` route parameter.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_SandboxDetailResponse"
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: GetSandbox handles `GET /api/v1/sbx/sandboxes/{sandbox_id}`.
  /api/v1/sbx/sandboxes/{sandbox_id}/connect:
    post:
      description: ConnectSandbox handles `POST /api/v1/sbx/sandboxes/{sandbox_id}/connect`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_ConnectSandbox
      parameters:
      - in: path
        name: sandbox_id
        required: true
        schema:
          description: SandboxID is bound from the `:sandbox_id` route parameter.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                timeout:
                  description: Timeout sets the sandbox lifetime in seconds after connecting; negative values are treated as 0.
                  format: int32
                  type: integer
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_SandboxResponse"
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: ConnectSandbox handles `POST /api/v1/sbx/sandboxes/{sandbox_id}/connect`.
  /api/v1/sbx/sandboxes/{sandbox_id}/logs:
    get:
      description: GetSandboxLogs handles `GET /api/v1/sbx/sandboxes/{sandbox_id}/logs`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_GetSandboxLogs
      parameters:
      - in: path
        name: sandbox_id
        required: true
        schema:
          description: SandboxID is bound from the `:sandbox_id` route parameter.
          type: string
      - in: query
        name: cursor
        schema:
          description: Cursor is the provider log cursor.
          format: int64
          type: integer
      - in: query
        name: direction
        schema:
          description: Direction controls log pagination direction.
          type: string
      - in: query
        name: level
        schema:
          description: Level filters logs by severity level.
          type: string
      - in: query
        name: limit
        schema:
          description: Limit is the maximum number of items to return.
          format: int32
          type: integer
      - in: query
        name: search
        schema:
          description: Search filters logs by content keyword.
          type: string
      - in: query
        name: start
        schema:
          description: Start specifies the log query start time or provider start cursor.
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_SandboxLogsResponse"
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: GetSandboxLogs handles `GET /api/v1/sbx/sandboxes/{sandbox_id}/logs`.
...
  /api/v1/sbx/sandboxes/{sandbox_id}/timeout:
    post:
      description: SetSandboxTimeout handles `POST /api/v1/sbx/sandboxes/{sandbox_id}/timeout`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_SetSandboxTimeout
      parameters:
      - in: path
        name: sandbox_id
        required: true
        schema:
          description: SandboxID is bound from the `:sandbox_id` route parameter.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                timeout:
                  description: Timeout sets the new sandbox lifetime in seconds; negative values are treated as 0.
                  format: int32
                  type: integer
              type: object
      responses:
        "200":
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: SetSandboxTimeout handles `POST /api/v1/sbx/sandboxes/{sandbox_id}/timeout`.
  /api/v1/sbx/templates:
    get:
      description: ListTemplates handles `GET /api/v1/sbx/templates`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_ListTemplates
      parameters:
      - in: query
        name: api_key_id
        schema:
          description: APIKeyID filters by the API key that created the resource; empty means no API key filter.
          type: string
      - in: query
        name: name
        schema:
          description: Name filters by template name.
          type: string
      - in: query
        name: build_status
        schema:
          description: BuildStatus filters by latest build status.
          type: string
      - in: query
        name: public
        schema:
          description: Public filters by visibility; empty means no visibility filter.
          type: string
      - in: query
        name: cursor
        schema:
          description: Cursor is the pagination cursor.
          type: string
      - in: query
        name: limit
        schema:
          description: Limit is the maximum number of items to return.
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/handler_TemplateResponse"
                type: array
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: ListTemplates handles `GET /api/v1/sbx/templates`.
    post:
      description: CreateTemplate handles `POST /api/v1/sbx/templates`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_CreateTemplate
      requestBody:
        content:
          application/json:
            schema:
              properties:
                alias:
                  description: Alias is the SDK-compatible template alias.
                  type: string
                cpu_count:
                  description: CPUCount is the default CPU count for the template.
                  format: int32
                  type: integer
                disk_size_mb:
                  description: DiskSizeMB is the default disk size in MB.
                  format: int32
                  type: integer
                dockerfile:
                  description: Dockerfile is the Dockerfile content used to build the template.
                  type: string
                memory_mb:
                  description: MemoryMB is the default memory size in MB.
                  format: int32
                  type: integer
                name:
                  description: Name is the template name; empty falls back to Alias.
                  type: string
                public:
                  description: Public indicates whether the template is publicly visible.
                  type: boolean
                ready_cmd:
                  description: ReadyCmd is the sandbox readiness check command.
                  type: string
                start_cmd:
                  description: StartCmd is the command executed after the sandbox starts.
                  type: string
              type: object
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_TemplateResponse"
          description: Accepted
        default:
          $ref: "#/components/responses/HTTPError"
      summary: CreateTemplate handles `POST /api/v1/sbx/templates`.
...
  /api/v1/sbx/templates/{template_id}:
    delete:
      description: DeleteTemplate handles `DELETE /api/v1/sbx/templates/{template_id}`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_DeleteTemplate
      parameters:
      - in: path
        name: template_id
        required: true
        schema:
          description: TemplateID is bound from the `:template_id` route parameter.
          type: string
      responses:
        "200":
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: DeleteTemplate handles `DELETE /api/v1/sbx/templates/{template_id}`.
    get:
      description: GetTemplate handles `GET /api/v1/sbx/templates/{template_id}`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_GetTemplate
      parameters:
      - in: path
        name: template_id
        required: true
        schema:
          description: TemplateID is bound from the `:template_id` route parameter.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_TemplateResponse"
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: GetTemplate handles `GET /api/v1/sbx/templates/{template_id}`.
    patch:
      description: UpdateTemplate handles `PATCH /api/v1/sbx/templates/{template_id}`.
      operationId: github_com_aonesuite_infra_internal_products_sandbox_handler_(_Module)_UpdateTemplate
      parameters:
      - in: path
        name: template_id
        required: true
        schema:
          description: TemplateID is bound from the `:template_id` route parameter.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                public:
                  description: Public updates template visibility; nil means no change.
                  type: boolean
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handler_UpdateTemplateResponse"
          description: OK
        default:
          $ref: "#/components/responses/HTTPError"
      summary: UpdateTemplate handles `PATCH /api/v1/sbx/templates/{template_id}`.
...